+++ CHANGELOG.md
@@ -3,10 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.2] - 2024-03-22
+## [0.1.2] - 2024-03-24
 ### Fixed
 - @javierfh03 - Fix incorrect exception when find nonexistent user.
 - @javierfh03 - Fix incorrect exception when find nonexistent group.
+- @javierfh03 - Fix incorrect command when remove user from group.
 ### Changed
 - @javierfh03 - Change group internal structure.